+++ bootmachine/__init__.py
@@ -1,2 +1,2 @@
 # following PEP 386
-__version__ = "0.5.9"
+__version__ = "0.6.0"