+++ tests/rackspacev2-salt-arch20132/settings.py
@@ -4,20 +4,20 @@ from bootmachine.settings_tests import *  # noqa
 
 
 CONFIGURATOR_MODULE = "bootmachine.contrib.configurators.salt"
-MASTER = "rackspacev2-salt-arch201208-a"
+MASTER = "rackspacev2-salt-arch20132-a"
 
 PROVIDER_MODULE = "bootmachine.contrib.providers.rackspace_openstack_v2"
 SERVERS = [
-    {"servername": "rackspacev2-salt-arch201208-a",
+    {"servername": "rackspacev2-salt-arch20132-a",
      "roles": ["salt-master"],
      "flavor": "2",
-     "image": "Arch 2012.08",
-     "distro_module": "bootmachine.contrib.distros.rackspace_arch_201208"},
-    {"servername": "rackspacev2-salt-arch201208-b",
+     "image": "Arch 2013.2",
+     "distro_module": "bootmachine.contrib.distros.rackspace_arch_20132"},
+    {"servername": "rackspacev2-salt-arch20132-b",
      "roles": ["test-additional-minions"],
      "flavor": "2",
-     "image": "Arch 2012.08",
-     "distro_module": "bootmachine.contrib.distros.rackspace_arch_201208"},
+     "image": "Arch 2013.2",
+     "distro_module": "bootmachine.contrib.distros.rackspace_arch_20132"},
 ]
 
 SSH_PUBLIC_KEY = os.path.join(os.environ["HOME"], ".ssh", "id_ecdsa.pub")